% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{rawNMRdata}
\alias{rawNMRdata}
\title{Raw 1D Bruker spectral data files from Salek \emph{et al.} (2007)}
\format{Raw 1D Bruker spectral data files or txt file}
\source{
\url{https://www.ebi.ac.uk/metabolights/MTBLS1}
}
\description{
This dataset is a subset of the public datasets from Salek *et al.* (2007).
The experiment has been designed to improve the understanding of early
stage of type 2 diabetes mellitus (T2DM) development. $^1$H-NMR human
metabolome was obtained from 25 healthy volunteers and 25 T2DM patients.
Raw 1D Bruker spectral data files were found in the MetaboLights database
(\url{https://www.ebi.ac.uk/metabolights/}, study MTBLS1).
}
\examples{
# Raw 1D Bruker spectral data files path
data_path <- system.file("extdata", "Human_diabetes_example",
                         package = "ASICSdata")

# txt file path
data_path <- system.file("extdata", "spectra_diabetes_example.txt",
                         package = "ASICSdata")
}
\references{
Salek, R. M., Maguire, M. L., Bentley, E., Rubtsov, D. V., Hough,
T., Cheeseman, M., ... & Connor, S. C. (2007). A metabolomic comparison of
urinary changes in type 2 diabetes in mouse, rat, and human.
\emph{Physiological genomics}, \strong{29}(2), 99-108.
}
\seealso{
\code{\link{designData}}
}
