% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{designData}
\alias{designData}
\title{Design experiment file from Salek \emph{et al.} (2007)}
\format{txt file}
\source{
\url{https://www.ebi.ac.uk/metabolights/MTBLS1}
}
\description{
This file contaning the design experiment of Salek \emph{et al.} (2007)
study.
}
\examples{
design_path <- system.file("extdata", "design_diabete_example.txt",
                           package = "ASICSdata")
}
\references{
Salek, R. M., Maguire, M. L., Bentley, E., Rubtsov, D. V., Hough,
T., Cheeseman, M., ... & Connor, S. C. (2007). A metabolomic comparison of
urinary changes in type 2 diabetes in mouse, rat, and human.
\emph{Physiological genomics}, \strong{29}(2), 99-108.
}
\seealso{
\code{\link{rawNMRdata}}
}
