% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{resASICS}
\alias{resASICS}
\title{Resultat of metabolite quantification performed with ASICS}
\format{txt file}
\description{
This dataset contains results of ASICS algorithm (Tardivel \emph{et al.},
2017) on spectra from Salek \emph{et al.} (2007).
}
\examples{
data_path <- system.file("extdata", "resuts_ASICS.txt",
                         package = "ASICSdata")
}
\references{
Salek, R. M., Maguire, M. L., Bentley, E., Rubtsov, D. V., Hough,
T., Cheeseman, M., ... & Connor, S. C. (2007). A metabolomic comparison of
urinary changes in type 2 diabetes in mouse, rat, and human.
\emph{Physiological genomics}, \strong{29}(2), 99-108.

Tardivel P., Canlet C., Lefort G., Tremblay-Franco M., Debrauwer
L., Concordet D., Servien R. (2017). ASICS: an automatic method for
identification and quantification of metabolites in complex 1D 1H NMR
spectra. \emph{Metabolomics}, \strong{13}(10): 109.
\url{https://doi.org/10.1007/s11306-017-1244-5}
}
